Chore: refactor replace.css
